refactor: Extract RDF parsing logic to ConstructResponseRdfDataParser
Extract the parsing phase of `ConstructResponseUtilV2` into a dedicated `ConstructResponseRdfDataParser` object. This separates concerns by moving all SPARQL result transformation logic (converting a `SparqlExtendedConstructResponse` into the `ConstructResponseRdfData` tree) into its own module, leaving the response building logic in `ConstructResponseUtilV2`.

The parser handles:
- Splitting resources from value/standoff data
- Building the resource and value hierarchy with permissions filtering
- Nesting dependent resources and incoming links
- All recursive traversal of the RDF graph
